Rename dir 'page' => 'pages
